pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "zen_creator"
version = "0.0.1"
description="ZEN-creator allows for the creation and management of datasets for ZEN-garden."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name="Jacob Mannhardt", email="zen-garden@ethz.ch" },
  { name="Christoph Funke"},
]

dependencies = [
    "numpy",  
    "pandas",
    "pydantic",
    "geopandas",
    "entsoe-py",
    "pyyaml",
    "tqdm",
    "pyarrow",
    "openpyxl",
    "eurostat",
    "scipy",
]

[project.optional-dependencies]

docs = [
    "sphinx==7.3.7",
    "myst-parser==3.0.1",
    "sphinx-book-theme==1.1.3",
    "pydata-sphinx-theme==0.15.4",
    "sphinx-reredirects==0.1.4",
    "nbsphinx==0.9.4",
    "nbsphinx-link==1.3.0",
    "sphinxcontrib-mermaid",
    "furo"
]

dev = [
    "black==25.12.0",
    "ruff==0.14.13",
    "mypy==1.19.1"
]

